Implement much more of the std::execution spec.
Implementing let_value requires having access to the
get_completion_signatures mechanics, so I added in most of the
building blocks needed to make this work.

Includes adding the beginnings of connect-time domain-based
customisation using the "late" form of transform_sender().

Replaced 'empty_env' with 'env<>' which is the new replacement.

Updated a lot of the concepts and CPOs to more accurately track
what is currently specified in the working draft and noted places
where the implementation deviates.

Added _get_tag() and _get_data() helpers to simplify detecting the
tag of a sender and unpacking a sender into its components.
These require implementing the tuple protocol explicitly as there
is no way to extract these automatically using structured bindings
with current compilers/stdlib facilities.